Tidy: split actions.templates.go into 2 files
